requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
configparser==3.5.1
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django~=2.1.7
django-debug-toolbar==1.9.1
django-summernote==0.8.11.4
djangorestframework==3.9.1
django-bootstrap-datepicker-plus
django-bootstrap4
entrypoints==0.3
enum34==1.1.6
geojson==2.4.1
gmaps==0.8.3
gunicorn==19.9.0
idna==2.8
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
MarkupSafe==1.1.0
mistune==0.8.4
pymongo
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
oauthlib==3.0.1
pandocfilters==1.4.2
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.5.0
prompt-toolkit==1.0.15
psycopg2==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.7.1
python-dateutil==2.7.5
python-openid==2.2.5
pytz==2018.9
pyzmq==17.1.2
redis==2.10.3
requests==2.21.0
requests-oauthlib==1.2.0
scandir==1.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wagtail
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.2
widgetsnbextension==3.4.2
django-filter
